+++ src/talkingheads/model_library/__init__.py
@@ -8,20 +8,6 @@ from .huggingchat import HuggingChatClient
 from .lechat import LeChatClient
 from .pi import PiClient
 
-
-def get_client(client_name):
-    """Returns the client by their tag name"""
-    return {
-        "ChatGPT": ChatGPTClient,
-        "Claude": ClaudeClient,
-        "Copilot": CopilotClient,
-        "Gemini": GeminiClient,
-        "HuggingChat": HuggingChatClient,
-        "LeChat": LeChatClient,
-        "Pi": PiClient,
-    }.get(client_name, None)
-
-
 __all__ = [
     "ChatGPTClient",
     "ClaudeClient",
